Create separate file for getLabelFromStackTrace tests
